# shared/api/openapi.yaml
openapi: 3.1.0
info:
  title: Catechesis Backend API
  description: |
    REST API for the Organized Catechesis system. Sprint 1 covers
    administrative endpoints (church and catechist provisioning) and
    class management for Lead Catechists.
  version: 0.1.0

servers:
  - url: http://localhost:8080
    description: Local dev

tags:
  - name: Admin
    description: Super-Admin operations — bootstrap parishes and Leads.

paths:
  /admin/churches:
    post:
      tags: [Admin]
      summary: Create a new Church (parish)
      description: |
        Bootstraps a new tenant. Only Super-Admins may invoke this
        endpoint. Returns the created Church.
      operationId: createChurch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChurchRequest'
      responses:
        '201':
          description: Church created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChurchResponse'
        '400':
          description: Validation error
        '403':
          description: Authenticated user is not a Super-Admin
  /admin/churches/{churchId}/catechists:
    post:
      tags: [Admin]
      summary: Create the first Lead Catechist for a Church
      description: |
        Bootstraps the first Lead for a given church. Only Super-Admins
        may invoke this endpoint. The created catechist always has role
        = LEAD. Attempting to create a second active Lead returns 409.
      operationId: createFirstLead
      parameters:
        - name: churchId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLeadRequest'
      responses:
        '201':
          description: Catechist created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatechistResponse'
        '400':
          description: Validation error
        '403':
          description: Authenticated user is not a Super-Admin
        '404':
          description: Church not found
        '409':
          description: Church already has an active Lead
  /classes:
    post:
      tags: [Classes]
      summary: Create a class in the Lead's own church
      description: |
        The authenticated Lead's church is used as the tenant for the
        new class. The request body has no churchId field — tenant
        identity is derived from authentication, not user input.
      operationId: createKlass
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateKlassRequest'
      responses:
        '201':
          description: Class created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KlassResponse'
        '400':
          description: Validation error
        '403':
          description: Authenticated user is not a Lead

components:
  schemas:
    CreateChurchRequest:
      type: object
      required: [displayName, timezone]
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 200
          example: "Paróquia Nossa Senhora de Fátima"
        timezone:
          type: string
          minLength: 1
          maxLength: 64
          example: "America/Sao_Paulo"

    ChurchResponse:
      type: object
      required: [id, displayName, timezone, createdAt]
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
        timezone:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    
    CreateLeadRequest:
      type: object
      required: [email, name]
      properties:
        email:
          type: string
          format: email
          minLength: 3
          maxLength: 200
          example: "marina@parish.org"
        name:
          type: string
          minLength: 1
          maxLength: 200
          example: "Marina Silva"

    CatechistResponse:
      type: object
      required: [id, churchId, email, name, role, active, createdAt]
      properties:
        id:
          type: string
          format: uuid
        churchId:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [LEAD, CATECHIST]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateKlassRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          example: "Sementinha III A"

    KlassResponse:
      type: object
      required: [id, churchId, name, active, createdAt]
      properties:
        id:
          type: string
          format: uuid
        churchId:
          type: string
          format: uuid
        name:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time